Correction du bug de prix lors de l'importation de bouteilles SAQ

+++ data/vinodb.sql
@@ -42,7 +42,7 @@ CREATE TABLE `vino__bouteille`
   `code_saq` VARCHAR(50) DEFAULT NULL,
   `pays` VARCHAR(50) DEFAULT NULL,
   `description` VARCHAR(200) DEFAULT NULL,
-  `prix_saq` FLOAT DEFAULT NULL,
+  `prix_saq` DECIMAL(6,2) DEFAULT NULL,
   `url_saq` VARCHAR(200) DEFAULT NULL,
   `format` VARCHAR(20) DEFAULT NULL,
   `id_type` INT DEFAULT NULL,
@@ -64,6 +64,7 @@ INSERT INTO `vino__bouteille` VALUES(NULL, 'Huber Riesling Engelsberg 2017', '//
 INSERT INTO `vino__bouteille` VALUES(NULL, 'Dominio de Tares Estay Castilla y Léon 2015', '//s7d9.scene7.com/is/image/SAQ/13802571_is?$saq-rech-prod-gril$', '13802571', 'Espagne', 'Vin rouge\r\n         \r\n      \r\n      \r\n      Espagne, 750 ml\r\n      \r\n      \r\n      Code SAQ : 13802571', 18, 'https://www.saq.com/page/fr/saqcom/vin-rouge/dominio-de-tares-estay-castilla-y-leon-2015/13802571', ' 750 ml', 1);
 INSERT INTO `vino__bouteille` VALUES(NULL, 'Tessellae Old Vines Côtes du Roussillon 2016', '//s7d9.scene7.com/is/image/SAQ/12216562_is?$saq-rech-prod-gril$', '12216562', 'France', 'Vin rouge\r\n         \r\n      \r\n      \r\n      France, 750 ml\r\n      \r\n      \r\n      Code SAQ : 12216562', 21, 'https://www.saq.com/page/fr/saqcom/vin-rouge/tessellae-old-vines-cotes-du-roussillon-2016/12216562', ' 750 ml', 1);
 INSERT INTO `vino__bouteille` VALUES(NULL, 'Tenuta Il Falchetto Bricco Paradiso -... 2015', '//s7d9.scene7.com/is/image/SAQ/13637422_is?$saq-rech-prod-gril$', '13637422', 'Italie', 'Vin rouge\r\n         \r\n      \r\n      \r\n      Italie, 750 ml\r\n      \r\n      \r\n      Code SAQ : 13637422', 34, 'https://www.saq.com/page/fr/saqcom/vin-rouge/tenuta-il-falchetto-bricco-paradiso---barbera-dasti-superiore-docg-2015/13637422', ' 750 ml', 1);
+
 
 
 -- -----------------------------------------------------
@@ -101,7 +102,7 @@ CREATE TABLE `vino__bouteille__collection`
   `date_achat` DATE DEFAULT NULL,
   `garde_jusqua` VARCHAR(200) DEFAULT NULL,
   `notes` VARCHAR(200) DEFAULT NULL,
-  `prix` FLOAT DEFAULT NULL,
+  `prix` DECIMAL(6,2) DEFAULT NULL,
   `quantite` INT DEFAULT NULL,
   `millesime` INT DEFAULT NULL,
   `id_usager` INT,
